--- src/main/java/base/AuthController.java
@@ -1,78 +1,0 @@
-package base;
-
-/**
- * Created by Marthxander on 5/10/2017.
- */
-import base.security.jwt.JwtAuthenticationRequest;
-import base.security.jwt.JwtTokenUtil;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.ResponseEntity;
-import org.springframework.mobile.device.Device;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.servlet.http.HttpServletRequest;
-
-@RestController
-public class AuthController {
-
-  @Value("${jwt.requestHeader}")
-  private String tokenHeader;
-
-  @Autowired
-  private AuthenticationManager authenticationManager;
-
-  @Autowired
-  private JwtTokenUtil jwtTokenUtil;
-
-  @Autowired
-  private UserDetailsService userDetailsService;
-
-  @RequestMapping(value = "${jwt.route.authentication.path}", method = RequestMethod.POST)
-  public ResponseEntity<?> createAuthenticationToken(@RequestBody JwtAuthenticationRequest authenticationRequest, Device device) throws AuthenticationException {
-
-    // Perform the security
-    final Authentication authentication = authenticationManager.authenticate(
-        new UsernamePasswordAuthenticationToken(
-            authenticationRequest.getUsername(),
-            authenticationRequest.getPassword()
-        )
-    );
-    SecurityContextHolder.getContext().setAuthentication(authentication);
-
-    // Reload password post-security so we can generate token
-    final UserDetails userDetails = userDetailsService.loadUserByUsername(authenticationRequest.getUsername());
-    final String token = jwtTokenUtil.generateToken(userDetails, device);
-
-    // Return the token
-    return ResponseEntity.ok(token);
-  }
-
-  @RequestMapping(value = "${jwt.route.authentication.refresh}", method = RequestMethod.GET)
-  public ResponseEntity<?> refreshAndGetAuthenticationToken(HttpServletRequest request) {
-    String token = request.getHeader(tokenHeader);
-    String username = jwtTokenUtil.getUsernameFromToken(token);
-    UserDetails user = userDetailsService.loadUserByUsername(username);
-
-    return ResponseEntity.ok(jwtTokenUtil.refreshToken(token));
-
-//        if (jwtTokenUtil.canTokenBeRefreshed(token, user.getLastPasswordResetDate())) {
-//            String refreshedToken = jwtTokenUtil.refreshToken(token);
-//            return ResponseEntity.ok(new JwtAuthenticationResponse(refreshedToken));
-//        } else {
-//            return ResponseEntity.badRequest().body(null);
-//        }
-  }
-
-}
