+++ src/main/java/base/security/config/WebSecurityConfig.java
@@ -1,0 +1,43 @@
+package base.security.config;
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.data.repository.query.SecurityEvaluationContextExtension;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@Configuration // Tell Spring this is a configuration class
+@EnableWebSecurity // Tell Spring to enable security as a web application
+@EnableGlobalMethodSecurity(prePostEnabled=true) // Enable @Pre @PostAuthorize
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter{
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.csrf().disable();
+
+        http
+                .authorizeRequests()
+                // allow anonymous resource requests
+                .antMatchers(HttpMethod.GET,"/resources/**").permitAll()
+                .antMatchers("/").permitAll() // to get auth token
+                .antMatchers("/**").permitAll() // to get auth token
+                .anyRequest().authenticated();
+
+        // Custom JWT based security filter -- Check for JWT tokens
+    }
+
+    @Bean
+    public SecurityEvaluationContextExtension securityEvaluationContextExtension() {
+        return new SecurityEvaluationContextExtension();
+    }
+}
