--- src/main/java/base/security/model/UserController.java
@@ -1,76 +1,0 @@
-package base.security.model;
-
-/**
- * Created by Marthxander on 5/8/2017.
- */
-import base.catalogDb.Course;
-import base.security.model.CurrentUser;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
-
-@RestController
-@RequestMapping("/user")
-public class UserController  {
-
-  @Autowired
-  private UserRepository userRepository;
-
-  // Return only logged in user
-  @GetMapping
-  public UserDetails getCurrentUser(@CurrentUser UserDetails currentUser) {
-    return currentUser;
-  }
-
-  @RequestMapping("/all")
-  public List<User> getUsers(@CurrentUser UserDetails currentUser) {
-    ArrayList<User> users = new ArrayList<>();
-    if (currentUser.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
-      userRepository.findAll().forEach(users::add);
-    }
-    return users;
-  }
-
-  @PostMapping
-  public User create(@Valid @RequestBody User reqUser) {
-    User user = new User();
-    user.setEmail(reqUser.getEmail());
-    user.setFirstName(reqUser.getFirstName());
-    user.setLastName(reqUser.getLastName());
-    user.setPassword(new BCryptPasswordEncoder().encode(reqUser.getPassword()));
-    user.setId(System.currentTimeMillis());
-    return userRepository.save(user);
-  }
-
-  @DeleteMapping("{id}")
-  public void delete(@PathVariable Long id) {
-    // ADMIN Route
-    userRepository.delete(id);
-  }
-
-  @PutMapping("{id}")
-  public User update(@PathVariable Long id, @RequestBody User reqUser) {
-    User user = userRepository.findOne(id);
-    if (user == null) {
-      return null;
-    } else {
-      user.setEmail(reqUser.getEmail());
-      user.setFirstName(reqUser.getFirstName());
-      user.setLastName(reqUser.getLastName());
-      user.setEmail(reqUser.getEmail());
-      user.setPassword(new BCryptPasswordEncoder().encode(reqUser.getPassword()));
-      return userRepository.save(user);
-    }
-  }
-}
