+++ src/main/java/base/user/User.java
@@ -1,0 +1,112 @@
+package base.user;
+
+/*
+ * Copyright 2002-2013 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import org.hibernate.validator.constraints.Email;
+import org.hibernate.validator.constraints.NotEmpty;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+import javax.persistence.*;
+import java.io.Serializable;
+
+/**
+ * Represents a user in our system.
+ *
+ * <p>
+ * In a real system use {@link PasswordEncoder} to ensure the password is secured
+ * properly. This demonstration does not address this due to time restrictions.
+ * </p>
+ *
+ * @author Rob Winch
+ */
+@Entity
+@Table(name="users") // 'user' is a keyword in Postgres
+public class User implements Serializable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    @NotEmpty(message = "First name is required.")
+    private String firstName;
+
+    @NotEmpty(message = "Last name is required.")
+    private String lastName;
+
+    @Email(message = "Please provide a valid email address.")
+    @NotEmpty(message = "Email is required.")
+    @Column(unique=true, nullable = false)
+    private String email;
+
+    @NotEmpty(message = "Password is required.")
+    private String password;
+
+    public User() {}
+
+    public User(User user) {
+        this.id = user.id;
+        this.firstName = user.firstName;
+        this.lastName = user.lastName;
+        this.email = user.email;
+        this.password = user.password;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    @Override
+    public String toString() {
+        return "User [userid=" + id + ", firstName=" + firstName + ", lastName=" + lastName + ", email="
+                + email + "]";
+    }
+}
