+++ src/main/java/base/user/UserRepository.java
@@ -1,10 +1,8 @@
-package base.security.model;
+package base.user;
 
-/**
- * Created by Marthxander on 5/8/2017.
- */
+
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepository extends CrudRepository<User, Long> {
-  User findByEmail(String email);
+    User findByEmail(String email);
 }
