+++ src/main/resources/static/pages/contact.html
@@ -1,7 +1,7 @@
 <body id="signUpPage">
 <h1 id="signUpTitle"> SLOCHART Account</h1>
 
-<div ng-controller="ContactController">
+<div ng-controller="contactController">
   <div class ="newAccountBody-div" >
     <h1 id="body"></h1>
     <input ng-model="user.firstName" placeholder="First Name" class="txt-input" />
